requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
tqdm>=4.0.0
python-dotenv>=0.21.0
